Eliminate `self.table.<thingy>.0.insert` in `function.rs`
Summary:
In this diff I continue eliminating the use of panic-prone and slightly
misleading low-level calls to `self.table.<thingy>.0.insert` to get an
`Idx<K>` in favor of tracking these `Idx` values directly in ways that
make it much easier to reason locally about why we know they are valid.

This is good because an invalid `Idx`, which is easy to create when we
make uncontrolled calls to `self.table.<thingy>.0.insert`, is basically
a dangling pointer that will panic if we try to store it in a binding
and use it in solve. It also makes the general architecture of the
bindings code easier to read, at least in my opinion.

Reviewed By: ndmitchell

Differential Revision: D75177418

fbshipit-source-id: 2be244185ab56b188f3caf47a86848080cffd2f8