Eliminate `self.table.types.0.insert` for `implicit_return`
Summary:
The `implicit_return` function needs to scan the body of a function and produce
a `Key` for all the "last statement" nodes it finds.

There is no way to verify that this scan will agree with bindings, but we've
carefully implemented them to agree - in other words, we can and do assume
that any "last statement" we get when we scan will match a `Key / Binding`
pair that we'll produce when we *bind* the body.

That still leaves an easy way for this to panic - we could potentially wind
up using these without binding the statements in the body (e.g. if we used
this in the context of `no_type_check`). This does not occcur.

This diff uses a clearly documented helper to avoid the direct access to
`self.table.types` (which I intend to hide from the rest of bindings as part
of a refactor to eliminate nondeterminism), and also includes discussion about
why this is safe / what invariants are required for it to be safe in the
doc comments.

------

This diff eliminates the last use of `self.table.<thingy>` in bindings
code, except in `bindings.rs` itself (where it is okay - that's the *right*
place to do low-level, fiddly index mangling)

It's probably the least important diff in the stack, in the sense that using
`self.table.types.0.insert` is really the right thing to do; it only really needs
to move to `bindings.rs` if I want to make `self.table` private (which I think
might be helpful); the only fundamentally important thing here is the doc
comment explaining conditions under which it's okay to use implicit returns.

Reviewed By: yangdanny97

Differential Revision: D75179984

fbshipit-source-id: 4c4528f96ee202b694b2d506130bd737f6b95321