Tighten up function index juggling for overload
Summary:
**Context on the next several diffs**

To facilite working on determinism in the presence of the way we
use `Var` for type inference, I need to introduce some discipline
around how we track the context of what binding we are currently
creating.

As I was looking at how to do this, I realized we have significant
technical debt that makes sense to knock out at the same time: there
are quite a few places where we are inserting a key without a binding.

This is an okay thing to do - we often do need to make use of a
key that might not yet be in bindings but we know *will* be (and in
fact, by the end of the determinism work we'll do this more often because
a lot of work will involve typing `Usage` keys to the binding in which
they are used) but I want to do it in a way where the rust interfaces are more
clearly indicating what is happening.

As a result, my first step in preparing for a refactor that will allow
us to get determinism is going to be locking down access to `self.table`,
so that we can create interfaces that state the intent of various operations
more clearly, and where we can explain why potentially-panic-prone code
is in fact respecting invariants that aren't obvious from local context.

**This diff**

As I worked through the bindings code to understand all the current ways
that `Idx<Key>` is created, I found one scenario where we fetch the binding
for a function index.

This has to change if we are going to lock down `self.table`, and after
squinting at the code I realized that the invariant is spread across about 50
lines of code, but we can actually wrap it all up in a single helper function that
juggles the function indices, which also makes it much easier to see and
comment how this relates to overloads and why we're confident that the existing
`unwrap()` won't panic.

This also makes the `function_def` helper simpler; one of the biggest sources
of remaining complexity was the fact that we had a bunch of mutable index vars,
but with a helper to handle the fiddly details this disappears from the higher-level
code.

Reviewed By: ndmitchell

Differential Revision: D75155904

fbshipit-source-id: 1a3460b05632bd5238400d3482111c4fac1df5de