Make it obvious what's going on in `Delete`
Summary:
As part of preparing to refactor bindings so that it tracks enough context
to meet our needs for determinism (which will require tying `Usage` keys to
the context in which they are used) did an audit o our use of `self.table`.

One of the patterns I want to eliminate on the way to tracking context is
direct access from most of the bindings phase to `self.table`, and this
will require eliminating the uses of `self.table.<variety>.0.insert(...)`
by either refactoring to better patterns or wrapping them in helpers.

This particular case is interesting: the existing code was pretty hard to track
because we had two identical `if let` blocks almost back-to-back which
obscured the reason the invariant works: the reality is that we know it's safe to
use this `Usage` because we just called `ensure_mutable_name` (which also hints
at the fix: we ought to make `ensure_mutable_name` *return* the resulting
`Idx<Key>`, I'll do that next).

After combining the blocks it's much clearer what is happening, and what our
next step of refactoring ought to be.

Reviewed By: yangdanny97

Differential Revision: D75157055

fbshipit-source-id: 7f1c71399a55da2efd0695749ab1afb69cc5bd3f