Rename `ClassIndex` to `ClassDefIndex`
Summary:
There's absolutely nothing wrong with `ClassIndex` as a name for this datatype
in principle.

But in practice, we wind up with quite a bit of code that has to have both one
of these *and* an `Idx<KeyClass>` in scope at the same time, and the natural
name for the `Idx<KeyClass>` if we follow conventions consistent with the rest
of the codebase is `class_idx`.

But this isn't great... even if `ClassIndex` is a good name for the datatype
in the abstract, having a variable `class_idx` and another one `class_index`
in the same function is pretty confusing.

Renaming the datatype to `ClassDefIndex` and the variable to `def_index` (the
fact that it's a class def index is always pretty obvious from the context)
solves this problem, enabling us to use more consistent names without as much
risk of getting them confused.

This isn't really a safety issue (the types will prevent us from mixing them
up) but helps a lot with readability in my opinion.

I am able, in this diff, to rename the two `class_idx_key` vars, which I used
to avoid confusion but were bothering me because of the inconsistency. I'm
planning to refactor in a way where this kind of name juggling will become
more important, hence the renaming work upfront.

Reviewed By: yangdanny97

Differential Revision: D75168492

fbshipit-source-id: fa6eb8bc2a6eb5da0b717202cedb7b18fd93ba79