Fix a potential panic waiting to take down the IDE
Summary:
I started working on the change implied by the previous diff (making
sure `ensure_.*` functions return an `Idx<Key>` so that we don't have
to resort to a confusing and not-particularly-safe `table.types.insert`
to get the already-inserted key) and I realized there's a pretty major
issue here.

Because we *failed* to return the `Idx<Key>` and instead relied on
unverified side-effects, we didn't notice that there's a case where
we actually fail to create a `Key::Usage` - the comment suggests that
at the time, we expected a type error tied to ruff error correction to
prevent us from ever wanting to read off the usage.

But this is generally pretty unsafe, and in fact the `Delete` code
before would definitely panic if the ruff parser were ever capable of
producing an empty identifier.

That's not good - as an IDE we should assume invalid syntax is common,
and therefore make sure that if we are counting on all Identifiers getting
a Usage they really do get one. I'm not certain whether it was possible to
hit the panic given the current error-recovering-parser behavior, but we
definitely should not rely on implementation details on pain of crashing the
LSP.

This will also unblock me making sure `ensure_mutable_name` passes up the
`Idx<Key>`, which is what I'll continue doing in the next diff.

Reviewed By: yangdanny97

Differential Revision: D75158225

fbshipit-source-id: b2b49637d64a37b890cc5403ee91ff27b08191af