Use `ClassIndices` in `expr.rs`'s `super()` code
Summary:
Just as in the previous diff, we're eliminating the need for a code
path where AST nodes analyzed inside of a class scope to use
low-level and potentially panic-prone `self.table.<thingy>.0.insert`
calls to fetch indexes that we know via scope information should exist
by just directly passing these indices through the scope.

Here we are eliminating a case in the analysis of `super()` calls; in
the process we're happily able to eliminate logic where, in spite of
having a `class_idx` in the first place, we were synthesizing a new
one and then re-hashing it.

Reviewed By: ndmitchell

Differential Revision: D75178435

fbshipit-source-id: 1acc929d325b29adee0a2c103aeb19f96fa3879f