Create `ClassIndices` and use it in `Scope`
Summary:
A big chunk of the tech debt cleanup I need to do is avoid the use
of `self.table.<thingy>.0.insert` as a way to *fetch* the `Idx<K>`
for some key we already have (or we know we are going to create);
this works if we never make mistakes, but it's a pretty error-prone
pattern that can and has led to many panics when our increasingly-complex
bindings logic doesn't follow invariants.

One of the big cases where this happens is in classes; a *lot* of things
inside of the body of a class (including potentially deeply nested in methods
and inner functions) can potentially need access to an identity of the current
class, and there are multiple ways of accessing the identity:
- the ClassDefIndex, which is a simple counter
- the `Idx<KeyClass>` used to get the `Class`
- the `Idx<KeyClassMetadata>` used to get the `ClassMetadata`
- the `Idx<KeyClassSynthesizedfields>` used to get synthesized fields

All of these ids *always* exist, and we can and do depend on this in a way where
Pyrefly will definitely panic if they don't all exist, but the existing code makes
it nontrivial to even see that they *do* all exist.

This diff introduces a `ClassIndices` datatype that both provides us a visible,
rust-type-level marker that all these things exist and also gives us a single
way to access them from nested scopes.

In the next diff, I'll get rid of multiple instances of
`self.table.<thingy>.0.insert` by making better use of the scopes.

-------

Moving to this pattern of inserting the key ahead of time if we might need back
references to it from data we create as part of binding is also going to
directly play into a notion of "what is the binding I'm currently producing"
that I'm intending to use as a core part of an algorithm to eliminate
nondeterminism - the idea is basically that we really need to create the
relevant Idx *before* we start binding things, because that's the only way
we'll be able to reliably tie the creation of a `Usage` key to the binding it
gets used in.

Reviewed By: ndmitchell

Differential Revision: D75175680

fbshipit-source-id: 9102cf89970c6859507ffa0dc37936fa08c14b91