don't show implicitly defined attribute error if it's inherited
Summary:
We show an error if we think an attribute is defined by a method that's not `__init__`. However, the place where the error happens doesn't know what's defined in a parent class.

If it's defined in a parent class then we're just assigning to the inherited attribute, not defining a new one. Maybe we shouldn't even make a binding here? Unfortunately we don't know that till we solve the binding.

fixes https://github.com/facebook/pyrefly/issues/231

Reviewed By: stroxler

Differential Revision: D75243279

fbshipit-source-id: 763214c499fa4fdff88d53dee13a5e5b2d79f9e8