Massage + rename `define_{nonlocal,global}_name` a bit
Summary:
I had to look at this a while before I really understood what was
happening here. This diff isn't a big change, but I think defining
the `binding` via the match-on-lookup and ending with *always*
doing an insert makes the logic slightly clearer (and two lines shorter)

In addition, `ensure` is actually an incorrect name for this operation,
if we're following conventions elsewhere: an "ensure" operation actually
means we are trying to make sure we have a `Usage` / `Forward` pair for a
name in some position in an expr; this operation is actually *creating*
the `Definition` node relied upon by static scope lookups, which is a
fundamentally different kind of operation. To that end, rename it.

Context on why this matters is that determinism requires computing a possible
reverse dependency when we see a usage, which means understanding the details
of name lookups is essential; since I got confused I think it's worth
putting in little readability wins on the way.

Reviewed By: ndmitchell

Differential Revision: D75236519

fbshipit-source-id: fb0ab23cf008258b5ae3ce6e7bc534c1319eee45