Make TParamInfo carry PreInferenceVariance instead of Variance
Summary:
Context:
In types.rs, we are converting TParamInfo to TParam.

For variance inference, we would like to have all type params related to a class as an input to the algorithm, so we can solve everything together.

Right now, we are converting from pre to post variance inference in many places. Why not make TParamInfo take pre-inference variance, and then when we convert TParamInfo to TParam, we infer variance and then populate the correct results in TParam, which carries post inference variance?

This required some refactors involving moving the tparam logic into the solver, which I validated with rchen152

Reviewed By: rchen152

Differential Revision: D75170988

fbshipit-source-id: 8a2746acc5907e83f4ad3681c0252ccbbe31e56a