Track another issue in mutable capture handling
Summary:
We currently *always* produce type errors on augmented assign,
which is incorrect.

But once we change that behavior, we will also need to make sure
that augemented assigns respect the type in the surrounding scope;
at first glance it might seem like this is always true, but augmented
assign can potentially change the type.

This is good to track in a test case so that future refactors don't
get it wrong.

Reviewed By: yangdanny97

Differential Revision: D75257057

fbshipit-source-id: 1dfcc4cca4561e25514f2c4af9efa8973c23bfc8