Break up nonlocal/global vs normal lookups
Summary:
Our handling of scope is too complicated.

Jia got stuck on this when working on a feature / bugfix during PyCon
sprints (I don't remember details anymore, but I do rememeber he spent
hours staring at it), and now I'm having a very hard time working out
how to track first-usage for determinism because that requires tracking
lookups, which are too complicated to understand.

I suspect that our scope handling may actually be buggy, it feels like
the name lookups are fundamentally wrong. But I'm hoping I can avoid
getting bogged down in that prior to working on determinism; what I
absolutely need at a minimum is a separation between "normal" lookups
and the nonlocal/global ones so that I can deal with less complexity.

That's what this diff does: I copy/pasted the the function, then
simplified until I could split it into two lookups, one that's *only*
used for mutable captures (nonlocal and global) and another that's
*only* used for normal lookups (regular and mutable).

In order to make nondeterminism work, I need to make normal lookups
track where a looked-up-name is used; getting rid of the mutable capture
lookup styles from the interface of the core lookup functions simplifies
the set of APIs I need to deal with in a way that will be a big help here.

Reviewed By: yangdanny97

Differential Revision: D75246854

fbshipit-source-id: f12343db5f3da6ee550e91484a83453435cc73c5