Simplify a test that was testing two separate things
Summary:
After the previous diff where I recorded a bug in `global` handling, I noticed
that the test above is sort of testing the same thing, but is also testing the
behavior of a global declaration that comes after a local define.

Since only one of those two behaviors is buggy (the one that I track in the new
test), we should simplify the pre-existing test so that it can be marked as passing;
it also makes sense to change the name.

Reviewed By: rchen152

Differential Revision: D75247806

fbshipit-source-id: 659417b493cd86f1906f8f0dd2490a0603c18e0d