Track two more bugs in mutable captures
Summary:
After squinting a while, I realized the primary mutation
test for mutable captures has two problems.

The first is that it isn't testing augmented assign, and actualy
it turns out that we *don't* allow this when we *should*.

The second is that it is testing `del`, but the behavior defined
is too lax; it's not possible to safely statically analyze code
using `del` on mutable captures; the Python runtime does allow
it but there's no way a type checker should; in my new test for
this I add code that type checks but would crash the system to
make it more obvious why we should be conservative.

Unfortunately this means our current behavior is exactly backward,
we allow `del` that we should not allow and we disallow augmented
assignment that we should allow.

To track this, I split the test into two tests, and marked both
of them as buggy (which is 4 tests because we currently have
global and nonlocal duplicated; it would be nice if we could simplify
this a bit but I'm not 100% sure how best to do that yet).

Reviewed By: yangdanny97

Differential Revision: D75249741

fbshipit-source-id: 5cce6cc1d36e4a0b510737917389dcd00c882699