Track a bug in global handling
Summary:
The current lookup logic for globals is extremely complex, following
the same "look up the scope stack" pattern used for other names. This
is both unnecesary and actually wrong: global means global, so we
should not complain on a `global x` and then allow an illegal mutation
downstream due to the `Any`, we should allow the `global` declaration
and complain on an incorrect mutation downstream.

Reviewed By: rchen152

Differential Revision: D75247807

fbshipit-source-id: e933a2e18a854b3b3b8a681c7c860b800fc45705