Eliminate the term "forward_lookup"
Summary:
The name `forward_lookup` never meant a ton to me; at one time, it was a common
lookup interface that was used on both `LegacyTParamsBuilder` and `BindingsBuilder`
to do lookups, and the "forward" in the name actually meant "create an extra
binding if you need to".

But these days, we no longer use a common interface, which means the one on
BindingsBuilder is just kind of a historical artifact; what's really confusing is
that it *looks* like the name "forward_lookup" was intended to mean "do a lookup
and create a `Binding::Forward`, but that's not really the history of the code.

There are only two callsites of the the `BindingsBuilder::forward_lookup`
method anyway; it's clearer to just do the lookup and create the forward manually.
And the one in `LegacyTParamBuilder` is, in my opinion, better renamed to
`intercept_lookup`; the name is intended to mean the same thing as "forward lookup"
but avoid the double meaning with regard to `Binding::Forward` (which is important
because it is not, in fact, necessarily going to produce a `Binding::Forward` - it
might produce a `Binding::CheckLegacyTParam` instead!)

Reviewed By: yangdanny97

Differential Revision: D75254669

fbshipit-source-id: a2aff153c3f018a92bdbaaf6bf46dc6ffff1371f