Track more issues with `global` / `nonlocal`
Summary:
I'm reasonably convinced that our current mechanism for `global` and `nonlocal`
does not really work properly.

While I've found a number of issues so far that could potentially be handled
by layering still more complexity into our already complex lookup logic,
this one requires an entirely new strategy for the bindings: the issue
is that it is fundamentally not right to put a `Definition` into the current
flow-sensitive typing context for mutable captures.

We can't do this because if we do, we'll treat unannotated mutable captures
just like unannotated locals. But by doing that, we allow arbitrary assignments,
which is entirely unsafe with regard to the parent scope analysis (or the
analysis of any other functions that share the same mutable capture).

I'm not certain whether I'll tackle this soon - I'm still hoping to find a way
to work on determinism without dealing with this issue yet, but the current
handling is making this challenging - but we definitely want to track the
problems and make sure we tackle them.

Reviewed By: yangdanny97

Differential Revision: D75250586

fbshipit-source-id: 0a59829ec642a6d69596ae8b08c324afbcb40609