Reorganize annotated attribute assignment slightly
Summary:
This small change makes a more significant upcoming change much easier to read
with a smaller diff.

Unfortunately it does force us to do one extra clone of the `attr`, but
as a rule `attr` is small (usually just a couple of names), and I'm not
currently seeing a great way to avoid this clone and refactor into the
data flow we need to track first usage of `Vars` created in `Definition`
bindings.

Reviewed By: ndmitchell

Differential Revision: D75295026

fbshipit-source-id: 95dda98ad052b88e8cf31e477efdb9305fba6c3c