Consolidate on `ann` as the argument to `make_binding`
Summary:
Our varnames were inconsistent in a couple places which was mildly
annoying; more importantly in a few places it was `k` which is a bad
name because it suggests "key", which sounds like it might mean
the primary key (i.e. the `Idx<Key>`) of the curent binding; the name `ann` is
much clearer since it correctly suggests this is specifically a
`KeyAnnotation`.

This matters more to me than it otherwise would because I'm going to need
at least some of these callsites to *actually* track the key of the current
binding, at which point the name `k` becomes a bigger problem.

Reviewed By: yangdanny97

Differential Revision: D75295029

fbshipit-source-id: 77f66001adb027a313130f218105ac042547c3b7