Pass ExprOrBinding to `bind_subscript_assign`
Summary:
It is a bug to call `make_binding` naively here, which guarantees
we will have a contextual typing bug.

I don't have time to fully fix the contextual typing bug right now,
but we can at least make the binding code correct so that it's entirely
a solver-side issue; this makes my life working on determinism easier
because it was weird and wrong for subscript vs attribute assignments
to be handling this differently.

Reviewed By: yangdanny97

Differential Revision: D75295031

fbshipit-source-id: 88d45ebd40ac52bdd5190e364f59a2a9c790de76