Lift TParamInfo to TParam conversion to class definition
Summary:
We would like to add a new binding for computing variance for a given type variable associated with a class.

The new binding would need to handle TParamInfo (which carries pre-inference variance) and convert it to TParam (which carries post-inference variance).

I think that we would need to delay converting TParamInfo -> TParam. So this refactor is the first step to doing this.

Reviewed By: rchen152

Differential Revision: D75326193

fbshipit-source-id: 01b2ceb055eb0c7328a2c6bb489d1979523c7687