Simplify `lookup_legacy_tparam` a bit more
Summary:
We don't have to use `break` and `continue`, which take already
confusing logic and make it even more confusing: we always give
up if we hit the fallback case, and we always do early return
otherwise *except* on `Forward`.

The logic is very slightly longer, but now we can describe the exit cases
better (one of them I don't fully understand, but at least it's good to know
that and describe it as best I can).

Reviewed By: rchen152

Differential Revision: D75333426

fbshipit-source-id: 253619dc9751b6b0dce1dd5fafe6da5c6147cd90