Simplify `insert_anywhere`
Summary:
I ran into a roadblock handling determinism: I *want* to be able to pass the
`Idx<Key>` of the binding we are currently constructing to `ensure_expr`, but
we rely heavily on callbacks in `bind_assign` and `bind_target`.

This causes significant complications, and I'm doing a pass of simplifications
(which are good BE work on their own, so worst case if I give up we'll still
be in a better state) as part of understanding my options.

In particular, I'm trying to see how tightly we can constrain the need for
mutations on `self.table`. The `insert_anywhere` handle was an egregious case
where we were using `&mut self` to return `&mut` data back to the caller, which
has cascading implications for the borrow checker, but we don't actually
need to - the only thing we ever do is insert one key into the map! By just passing
it into `insert_anywhere`, we can make this a simple unit-valued function.

Reviewed By: rchen152

Differential Revision: D75333423

fbshipit-source-id: 8bf771c068d7c905fc0bc14105cd2c981e6e90b7