Add another failing scope test case
Summary:
Here, we are failing to detect that augmented assignment and deletion actually
both define a local (as in, the python compiler will create the slot for it,
and attempt to use that slot on operations - it will *not* try to do anything
to a non-mutable capture because the presence of the local prevents capture),
so we are confusing illegal mutation vs uninitialized local.

Either way we get a type error, but it's the wrong type error and
this hints at how any refactor of scope lookups ought to work.

In the case where there's no shadowing, we get reasonable errors, the holes
in our analysis are only visible to users in the presence of shadowing.

Reviewed By: yangdanny97

Differential Revision: D75359385

fbshipit-source-id: dc716edc672206e0c078405683fd8f5073c39c42