Don't crash when producing completions (#355)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/355

A many-step reproduce required. We were checking for completions, which then does a type check of the attributes. That type check uses no error collector, the current module, but a location from where the attribute was defined. Mixing modules and ranges from different places goes wrong. In this case it goes wrong only when we add an error into the collector, which only happens if we have an overloaded property, as overload passes a real error collector.

Solution is just use the empty range if we don't want errors, as the empty range is valid for all modules.

Fixes #347

Reviewed By: SamChou19815

Differential Revision: D75373926

fbshipit-source-id: 5fe5ccee37a70340782601af9ba29f6f9d3c5aaf