Explain what's going on a bit better in `If`
Summary:
I had to squint at this a bit to work out what's going on (partly because the
current code isn't as clear as it could be, I'm going to try to make it clearer
and reduce the amount of potentially unsafe mutation ... `mem::take` puts
`self.scopes` in a temporarily invalid state so I'm hoping to refactor away
from using it).

Let's add a comment to make it more obvious before I start messing with the logic.

Reviewed By: yangdanny97

Differential Revision: D75361088

fbshipit-source-id: 26a11a4514158fd5f3af58b4a223a11e0744dc9b