Make `setup_loop` make more sense
Summary:
It's weird to increment the loop depth at the top but only push
the loop at the bottom.

It can be done in one place: it has to be done after we've generated
the phi keys for the loop-back, and before we bind narrow ops.

This makes it so there's actually a clear point in the rust control
flow where we "enter the loop"; before we were in a halfway state
for a couple of operations which was fishy.

Reviewed By: yangdanny97

Differential Revision: D75361093

fbshipit-source-id: 89111193b9b0ed889e54654717b2e9c33a936670