Eliminate unnecessary `no_next` mutation
Summary:
We introduced this when we were optimizing `insert_phi_keys`,
because the original created a *new* flow with `no_next` hardcoded
to `false`.

But there's no actual need for this, the loop entry can certainly inherit
whatever the base flow has, and it's a little confusing to have an unnecessary
mutation in there because it looks like it must be important somehow.

Reviewed By: yangdanny97

Differential Revision: D75361096

fbshipit-source-id: a5abd776806d44a504fc3458ff087fa2ff85de7d