Rename `no_next -> has_terminated`, improve comment
Summary:
In the name "no_next", the meaning written in plain English is this:
"Should the current flow be included inside the merge of itself and
all of its sibling, and possibly the base flow, when we are ready to
merge them after this flow and all of its siblings have been analyzed?"

That's a mouthful, and is a hint that the name is too coupled to how
it is *used* (in one of the most complicated parts of bindings, the flow
merging), rather than what it actually *represents* ("have we seen control
flow terminate already this flow?").

Rename it; it's still good to discuss how it is used, but we can do that
in the doc comment.

Reviewed By: yangdanny97

Differential Revision: D75361087

fbshipit-source-id: f909f31f0aa76b06c33068dc0f5a87aa700a0eb8