Don't negate if branches if exhaustive
Summary:
One reason I wass having trouble fully understanding what happens
in `If` is that there were junk operations running: if the `If`
is exhaustive, we're going to throw away the current flow info, so
it doesn't make any sense to negate the ops... the only thing we
are doing is wasting both RAM and compute binding them and later
solving them even though they are unreachable.

It's both clearer and more efficient to move this into the non-exhaustive
block.

Reviewed By: yangdanny97

Differential Revision: D75361094

fbshipit-source-id: 3d34478b4ff519c1cabcd01c780631afb72aaa0a