Move `loop_depth` to `Scopes`
Summary:
It is pretty strange to duplicate the definition of loop
depth, we get a lot of extra passing of data that isn't needed
and more chance of things going out of sync.

It also forces more references to be mutable than necessary, which is
causing major architectural issues for me working on determinism.

Reviewed By: yangdanny97

Differential Revision: D75355660

fbshipit-source-id: 0f0697850e9467491aaf6bf0e3e2ded62ef0bd64