Use a helper to pop and decompose function scopes
Summary:
The scope popping in `function_body_scope` is a bit fiddly, and also was
missing a check that this is actually a function scope (which it always is, but
it's nice to check because if it ever didn't we would want to panic with a nice
error message rather than continuing on with nonsense in our scope stack).

Let's move it to `Scopes`. I am motivated partly by noting that the
function_yields_and_returns is sort of behaving like a parallel scope stack
(although it only lines up with a subset of the scopes, so they aren't exactly
the same), and creating smarter helpers for dealing with function scope will
give me an obvious way to tie the two stacks together.

Reviewed By: ndmitchell

Differential Revision: D75381940

fbshipit-source-id: 323c558225dd0c4f4efbf186b6ae5b899cc37a74