Rearrange lambda binding a bit
Summary:
We need to handle yields in lambda (in the long term, we need it to
be correct; in the short term I at least need it to not panic).

To that end, we shouldn't rely on the vanilla scope handling in
`ensure_expr` because we need to actively manage the scope.

Reviewed By: ndmitchell

Differential Revision: D75409589

fbshipit-source-id: 504d8073b7a4ebb3504196adebb291672dd14543