Eliminate top-level `function_yields_and_returns`
Summary:
The yield and return handling in BindingsBuilder is pretty messy - we
blindly store yields and returns using unmanaged mutation, and then
check them at the very end (paying some cost in ram along the way, and
also limiting our options for refactoring).

It's much nicer to just detect when we are at the top-level and handle
right away. Some helpers make this easy to do, and also prevents
`bindings.rs` from having to deal with logic that is better handled
in `expr.rs` and `stmt.rs`.

Reviewed By: ndmitchell

Differential Revision: D75381943

fbshipit-source-id: e24bbca5d04b025283839c833d944cda9b094b70