Move flow style merge to `FlowStyle::merged`
Summary:
This function was taking an `&mut BindingsBuilder`, and never
using it! It's much clearer to have a pure helper function as a static
function on `FlowStyle`.

I prefer "merged" to "merge" because the former name implies a pure
function *returning* a flow, the latter implies mutation of something
passed in.

Reviewed By: ndmitchell

Differential Revision: D75388296

fbshipit-source-id: 9b06305bb9a3626b509c32a006ca4a3cf8303fdf