Use a helper for "static ann override logic"
Summary:
I suspect that this operation won't exist once we change mutable capture
handling to fix bugs (trying to just pull the outer scope *annotation* in
doesn't really work - e.g. it allows arbitrary mutations of unannotated
mutable captures which is a pretty major bug).

That said, to unblock refactors for determinism I am working to use
only well-controlled mutations of `&mut scope` values; as a result, I'm
making a helper for this operation as-is because I don't have time to rethink
mutable captures at the moment.

Reviewed By: ndmitchell

Differential Revision: D75388298

fbshipit-source-id: 83635afaa5a26450a0414bb5a75d0f30106a296f