Further simplify special export lookups
Summary:
The previous diff removed a layer of indirection in `SpecialExport`
handling by getting rid of the `SpecialEnv` trait.

This diff further simplifies by getting rid of `SpecialEntry`, which
isn't needed: we have all the information we need in `BindingsBuilder`
to directly pass down enough for `Scopes` to fully resolve special
exports: all we need to know is whether we're resolving a bare name
or an attribute on a name.

The logic still isn't trivial, but now it's a straightforward two-stage
computation in two layers rather relying on an extra datatype.

Reviewed By: ndmitchell

Differential Revision: D75388299

fbshipit-source-id: 66e4c0fb349b4f99830ca0622e4066e98d02f3fd