Create `set_current_flow_to_merged_branches` helper
Summary:
It's much easier to work with and refactor flow operations that are mutating
both bindings and scope if we centralize the handling into well-named helper
functions in `bindings.rs`.

The new name also reflects the semantics better I think: we aren't "merging"
any existing active flow, we are instead creating a brand-new "merged" flow
from the branches and setting current to it (discarding the current flow
in the process). Careful consideration of nouns and verbs in the name makes
it a little clearer what operation is being peformed, and how it differs
from `merge_branches_into_current` which is actually merging current as well.

Reviewed By: ndmitchell

Differential Revision: D75388294

fbshipit-source-id: 604bed72c9ff34d159c0346d56a494cf81226c70