Clean up `negate_and_merge_flow`
Summary:
The `negate_and_merge_flow` function in `expr.rs` is the last place
we have an un-managed `mem::take` of the current flow. This is not
desirable: the state of `Scopes` is invalid at the end of such an
operation, if possible it would be better to always swap out the flow
with a valid one.

This is easiest to do efficiently by making `set_current_flow` return the
old flow; many callsites will ignore it but the negate_and_merge_current_flow
will not.

The resulting code is both safer (in that `Scopes` doesn't enter a temporary
invalid state) and more readable because we don't need `take` and `current_mut`.

NOTE: There might be a small efficiency loss here; `mem::take` is cheap but not
free. If this is a perf concern, we could avoid unnecessary `take` in cases
where the caller doesn't care introducing a `swap_current_flow` function to do
the new operation while letting `set_current_flow` continue to overwrite and
return unit. My instinct was that this wasn't worth an extra method, but happy
to reconsider.

Reviewed By: ndmitchell

Differential Revision: D75388292

fbshipit-source-id: d13bafc9fff9844940024186b8164f3e839ba968