Simplify special env lookups
Summary:
I need to make a change to how the `BindingsBuilder` and
`Scopes` interact here, but the existing code is so complicated it's
hard to even read. None of this indirection is needed, we can just use
a straightforward `as_special_export` method.

I suspect that the indirection made sense at some point in the past,
we must have had two different implementations of `SpecialEnv` but
we no longer do.

Reviewed By: ndmitchell

Differential Revision: D75388295

fbshipit-source-id: 10d84a9ca7a0f6d0b1a1a0a606c93614a1a1f7a7