Move function yield and return storage into `Scopes`
Summary:
The yields and returns are naturally aligned with scopes - not all
scopes have them, but a reverse search for the innermost function
scope will find the relevant one, and if we find nothing then
we are looking at a top-level yield or return and we should handle
it accordingly.

This eliminates duplicated data that had to be kept in sync manually,
and also helps control mutations in a more modular way.

There's a slightly funny pattern here that lets us avoid unnecessary clones
where we pass ownership of the return / yield / yield from, and then pass it
*back* if it is rejected. I use the name `oops_top_level` in callers for the
rejected operation, which I think makes it pretty easy to follow what's going
on here even though it's a fiddly rust-specific optimization.

Reviewed By: ndmitchell

Differential Revision: D75381942

fbshipit-source-id: 05e257c640b32512f91fceb933c7b770f169833c